+++ app/components/timeline/timeline-service.js
@@ -123,13 +123,13 @@ angular.module('lizard-nxt')
             this.drawAggWindow
           );
         }
-        this._svg.on("click", clicked);
+        this._svg.select('#listeners').on("click", clicked);
       }
     },
 
     removeClickListener: {
       value: function () {
-        this._svg.on("click", null);
+        this._svg.select('#listeners').on("click", null);
       }
     },
 
@@ -163,7 +163,7 @@ angular.module('lizard-nxt')
 
         if (!aggWindow) {
           height = this._getHeight(this.dimensions);
-          aggWindow = this._svg.append("g")
+          aggWindow = this._svg.select('g').append("g")
             .attr('class', 'agg-window-group');
           aggWindow
             .append("rect")
@@ -373,11 +373,15 @@ angular.module('lizard-nxt')
         if (zoomEndFn) {
           zoomend = setZoomEndFunction(zoomEndFn);
         }
-        this._svg.call(d3.behavior.zoom()
+        this._svg.select('#listeners').call(d3.behavior.zoom()
           .x(xScale)
           .on("zoom", zoomed)
           .on("zoomend", zoomend)
         );
+
+        // Move listener rectangle to the front
+        var el = this._svg.select('#listeners').node();
+        el.parentNode.appendChild(el);
       }
     }
   });
@@ -416,6 +420,50 @@ angular.module('lizard-nxt')
     var axisEl = svg.select('#xaxis')
         .attr("class", "x axis timeline-axis");
 
+    addClickToAxisTicks(axisEl.selectAll('text'));
+  };
+
+  /**
+   * Takes a d3 multiselection of text elements and add click interaction to
+   * zoom to rounded dates.
+   * @param {d3 selections} ticks text elements of tick marks.
+   */
+  var addClickToAxisTicks = function(ticks) {
+    ticks
+      .each(function (d) {
+        if (d.getMinutes() === 0) {
+          d3.select(this).attr('class', 'clickable');
+        }
+      })
+      .on('click', zoomToHourDayMonthOrYear);
+  };
+
+  /**
+   * Gets a date object, typically from a d3 tick mark. If it is a round
+   * year|month|day|hour it zooms the timeline and calls zoom callbacks. It does
+   * not zoom to minutes and seconds.
+   * @param  {date} d
+   */
+  var zoomToHourDayMonthOrYear = function (d) {
+    var end = new Date(d.getTime());
+    if (d.getHours() === 0) {
+      if (d.getDate() === 1) {
+        if (d.getMonth() === 0) {
+          xScale.domain([d, end.setYear(d.getFullYear() + 1)]);
+        }
+        else {
+          xScale.domain([d, end.setMonth(d.getMonth() + 1)]);
+        }
+      }
+      else {
+        xScale.domain([d, end.setDate(d.getDate() + 1)]);
+      }
+    }
+    else {
+      xScale.domain([d, end.setHours(d.getHours() + 1)]);
+    }
+    zoomed();
+    zoomend();
   };
 
   /**
@@ -565,7 +613,10 @@ angular.module('lizard-nxt')
   var setZoomFunction = function (
     svg, dimensions, xScale, zoomFn) {
     var zoomed = function () {
-      d3.event.sourceEvent.preventDefault();
+      // might not exist when called programmatically
+      if (d3.event.sourceEvent) {
+        d3.event.sourceEvent.preventDefault();
+      }
 
       var ONE_HOUR = 1000 * 60 * 60;
 
@@ -645,7 +696,10 @@ angular.module('lizard-nxt')
     var clicked = function () {
       // Check whether user is dragging instead of clicking
       if (!d3.event.defaultPrevented) {
-        var ts = xScale.invert(event.offsetX - dimensions.padding.left)
+        var x = d3.event.clientX
+          - UtilService.TIMELINE_LEFT_MARGIN
+          - dimensions.padding.left;
+        var ts = xScale.invert(x);
         clickFn(ts, dimensions);
       }
     };