+++ app/lizard-nxt-master-controller.js
@@ -73,40 +73,6 @@ angular.module('lizard-nxt')
 
   // END CONTEXT
 
-  // KEYPRESS
-
-  // If escape is pressed close box
-  // NOTE: This fires the watches too often
-  $scope.keyPress = function ($event) {
-
-    if ($event.target.nodeName === "INPUT" &&
-      ($event.which !== 27 && $event.which !== 13
-       && $event.which !== 32)) {
-      return;
-    }
-
-    // pressed ESC
-    if ($event.which === 27) {
-
-      State.box.type = "point";
-      State.spatial.here = undefined;
-      State.spatial.points = [];
-      ClickFeedbackService.emptyClickLayer(MapService);
-
-      // This does NOT work, thnx to Angular scoping:
-      // $scope.geoquery = "";
-      //
-      // ...circumventing-angular-weirdness teknique:
-      document.querySelector("#searchboxinput").value = "";
-    }
-    // TODO: move
-    // // play pause timeline with Space.
-    // if ($event.which === 32) {
-    //   $scope.timeState.playPauseAnimation();
-    // }
-
-  };
-
   $scope.toggleVersionVisibility = function () {
     $('.navbar-version').toggle();
   };